# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  api.Settings:
    properties:
      system:
        $ref: '#/definitions/api.SystemSettings'
      trading:
        $ref: '#/definitions/api.TradingSettings'
    type: object
...
  api.SystemSettings:
    properties:
      cacheSize:
        type: string
      debugMode:
        type: boolean
      logLevel:
        type: string
    type: object
  api.TradingSettings:
    properties:
      defaultStopLoss:
        type: integer
      dryRunMode:
        type: boolean
      maxPositionRatio:
        type: integer
      riskControl:
        type: boolean
    type: object
...
host: localhost:8082
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Quantitative Contract Automated Trading System API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: QCAT API
  version: "1.0"
paths:
...
  /settings:
    get:
      consumes:
      - application/json
      description: Get current system settings and configuration
      produces:
      - application/json
      responses:
        "200":
          description: Current system settings
          schema:
            $ref: '#/definitions/api.Settings'
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Get system settings
      tags:
      - Settings
    put:
      consumes:
      - application/json
      description: Update system settings and configuration
      parameters:
      - description: Updated settings
        in: body
        name: settings
        required: true
        schema:
          $ref: '#/definitions/api.Settings'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              message:
                type: string
              settings:
                $ref: '#/definitions/api.Settings'
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Update system settings
      tags:
      - Settings
...
  /ws/alerts:
    get:
      consumes:
      - application/json
      description: Establish WebSocket connection for real-time alerts and notifications
      produces:
      - application/json
      responses:
        "101":
          description: Switching Protocols
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Alerts WebSocket stream
      tags:
      - WebSocket
  /ws/market/{symbol}:
    get:
      consumes:
      - application/json
      description: Establish WebSocket connection for real-time market data streaming
      parameters:
      - description: Trading symbol
        in: path
        name: symbol
        required: true
        type: string
      produces:
      - application/json
      responses:
        "101":
          description: Switching Protocols
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Market data WebSocket stream
      tags:
      - WebSocket
  /ws/strategy/{id}:
    get:
      consumes:
      - application/json
      description: Establish WebSocket connection for real-time strategy status updates
      parameters:
      - description: Strategy ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "101":
          description: Switching Protocols
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Strategy status WebSocket stream
      tags:
      - WebSocket
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"